core/compiler core reformat and tidy
git-svn-id: https://svn.jboss.org/repos/labs/labs/jbossrules/trunk@10952 c60d74c8-e8f6-0310-9e8f-d4a2fc68ab70

+++ drools-core/src/test/java/org/drools/base/extractors/CharClassFieldExtractorTest.java
@@ -63,7 +63,7 @@ public class CharClassFieldExtractorTest extends BaseClassFieldExtractorsTest {
     public void testGetIntValue() {
         try {
             Assert.assertEquals( 'a',
-                                 (int) this.extractor.getIntValue( this.bean ) );
+                                 this.extractor.getIntValue( this.bean ) );
         } catch ( final Exception e ) {
             fail( "Should not throw exception" );
         }
@@ -90,7 +90,7 @@ public class CharClassFieldExtractorTest extends BaseClassFieldExtractorsTest {
     public void testGetValue() {
         try {
             Assert.assertEquals( 'a',
-                                 ((Number)this.extractor.getValue( this.bean )).intValue() );
+                                 ((Number) this.extractor.getValue( this.bean )).intValue() );
         } catch ( final Exception e ) {
             fail( "Should not throw an exception" );
         }
